+++ packages/terra-popup/src/Popup.jsx
@@ -105,6 +105,11 @@ const propTypes = {
    * Callback function to handle click events on the popup.
    */
   onClick: PropTypes.func,
+  /**
+   * Determines the positioning strategy for the Popup when used with Hookshot.
+   * @type {string}
+   */
+  hookshotContentPosition: PropTypes.string,
 };
 
 const defaultProps = {
@@ -121,6 +126,7 @@ const defaultProps = {
   isHeaderDisabled: false,
   isOpen: false,
   popupContentRole: 'dialog',
+  hookshotContentPosition: '',
 };
 
 class Popup extends React.Component {
@@ -216,7 +222,7 @@ class Popup extends React.Component {
     }
   }
 
-  createPopupContent(boundingFrame, showArrow) {
+  createPopupContent(boundingFrame, showArrow, hookshotContentPosition) {
     const boundsProps = {
       contentHeight: PopupHeights[this.props.contentHeight] || PopupHeights['80'],
       contentWidth: PopupWidths[this.props.contentWidth] || PopupWidths['240'],
@@ -257,6 +263,7 @@ class Popup extends React.Component {
         isWidthAutomatic={this.props.contentWidth === 'auto'}
         isFocusedDisabled={this.props.isContentFocusDisabled}
         onClick={this.props.onClick}
+        hookshotContentPosition={hookshotContentPosition}
       >
         {this.props.children}
       </PopupContent>
@@ -282,6 +289,7 @@ class Popup extends React.Component {
       onRequestClose,
       targetRef,
       targetAttachment,
+      hookshotContentPosition,
     } = this.props;
     /* eslint-enable no-unused-vars */
 
@@ -302,7 +310,7 @@ class Popup extends React.Component {
     if (showArrow) {
       cOffset = PopupUtils.getContentOffset(cAttachment, tAttachment, this.props.targetRef(), PopupArrow.Opts.arrowSize, cornerSize);
     }
-    const hookshotContent = this.createPopupContent(boundingRef ? boundingRef() : undefined, showArrow);
+    const hookshotContent = this.createPopupContent(boundingRef ? boundingRef() : undefined, showArrow, hookshotContentPosition);
 
     return (
       <React.Fragment>