+++ packages/terra-popup/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * `hookshotContentPosition` prop to control positioning strategy for the popup and popup content when used with Hookshot.
+
 ## 6.77.0 - (March 8, 2024)
 
 * Changed